+++ lib/Repository/ClubException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace OCA\SPGVerein\Repository;
+
+class ClubException extends \Exception { 
+
+}