+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## 0.8.9 – 2020-12-20
 ### Added
 - Nextcloud 20 compatibility (#12)
 - Add changelog information to distribution (#14)